rename SeqPattern0 to SeqPattern
