+++ core/src/test/scala/org/bykn/bosatsu/pattern/SeqPatternTest.scala
@@ -5,8 +5,15 @@ import org.scalacheck.{Arbitrary, Gen, Shrink}
 import org.scalatest.prop.PropertyChecks.{ forAll, PropertyCheckConfiguration }
 import org.scalatest.FunSuite
 
+import SeqPattern.{Cat, Empty}
+
+import cats.implicits._
+
 object StringSeqPatternGen {
-  import StringSeqPattern._
+  import SeqPart.{Wildcard, AnyElem, Lit}
+
+  type Named = NamedSeqPattern[Char]
+  val Named = NamedSeqPattern
 
   // generate a string of 0s and 1s to make matches more likely
   val genBitString: Gen[String] =
@@ -16,13 +23,16 @@ object StringSeqPatternGen {
       list <- Gen.listOfN(sz, g)
     } yield list.mkString
 
-  val genPart: Gen[Part] =
+  val genPart: Gen[SeqPart[Char]] = {
+    import SeqPart._
+
     Gen.frequency(
       (15, Gen.oneOf(Lit('0'), Lit('1'))),
       (2, Gen.const(AnyElem)),
       (1, Gen.const(Wildcard)))
+  }
 
-  val genPat: Gen[Pattern] = {
+  val genPat: Gen[SeqPattern[Char]] = {
 
     val cat = for {
       h <- genPart
@@ -34,9 +44,9 @@ object StringSeqPatternGen {
       (5, cat))
   }
 
-  implicit val arbPattern: Arbitrary[Pattern] = Arbitrary(genPat)
+  implicit val arbPattern: Arbitrary[SeqPattern[Char]] = Arbitrary(genPat)
 
-  implicit lazy val shrinkPat: Shrink[Pattern] =
+  implicit lazy val shrinkPat: Shrink[SeqPattern[Char]] =
     Shrink {
       case Empty => Empty #:: Stream.empty
       case Cat(Wildcard, t) =>
@@ -48,7 +58,7 @@ object StringSeqPatternGen {
   def genNamedFn(nextId: Int): Gen[(Int, Named)] = {
     lazy val recur = Gen.lzy(res)
 
-    lazy val genNm: Gen[(Int, Named.Bind)] =
+    lazy val genNm: Gen[(Int, Named.Bind[Char])] =
       recur.map { case (i, n) => (i + 1, Named.Bind(i.toString, n)) }
 
     // expected length = L = p_c * 2 * L + pn + L + pe * 0 + pp
@@ -57,61 +67,61 @@ object StringSeqPatternGen {
     // so we need 2*pc + pn < 1
     // e.g. pc = 1/3, pn = 1/9 would work
     // L = (4/9) / (1 - (2/3 + 1/9)) = 4 / (9 - 5) = 1
-    lazy val res: Gen[(Int, Named)] =
+    lazy val res: Gen[(Int, NamedSeqPattern[Char])] =
       Gen.frequency(
         (3, for {
           (i0, n0) <- recur
           (i1, n1) <- genNamedFn(i0)
-        } yield (i1, Named.NCat(n0, n1))),
+        } yield (i1, NamedSeqPattern.NCat(n0, n1))),
         (1, genNm),
-        (1, Gen.const((nextId, Named.NEmpty))),
-        (4, genPart.map { p => (nextId, Named.NPart(p)) }))
+        (1, Gen.const((nextId, NamedSeqPattern.NEmpty))),
+        (4, genPart.map { p => (nextId, NamedSeqPattern.NSeqPart(p)) }))
 
     res
   }
 
-  val genNamed: Gen[Named] =
+  val genNamed: Gen[NamedSeqPattern[Char]] =
     genNamedFn(0).map(_._2)
 
-  implicit val arbNamed: Arbitrary[Named] = Arbitrary(genNamed)
+  implicit val arbNamed: Arbitrary[NamedSeqPattern[Char]] = Arbitrary(genNamed)
 
   def interleave[A](s1: Stream[A], s2: Stream[A]): Stream[A] =
     if (s1.isEmpty) s2 else if (s2.isEmpty) s1 else {
       s1.head #:: interleave(s2, s1.tail)
     }
 
-  implicit val shrinkNamed: Shrink[Named] =
+  implicit val shrinkNamedSeqPattern: Shrink[NamedSeqPattern[Char]] =
     Shrink {
-      case Named.NEmpty => Stream.Empty
-      case Named.Bind(n, p) =>
-        val sp = shrinkNamed.shrink(p)
-        val binded = sp.map(Named.Bind(n, _))
+      case NamedSeqPattern.NEmpty => Stream.Empty
+      case NamedSeqPattern.Bind(n, p) =>
+        val sp = shrinkNamedSeqPattern.shrink(p)
+        val binded = sp.map(NamedSeqPattern.Bind(n, _))
         interleave(sp, binded)
-      case Named.NPart(p) =>
+      case NamedSeqPattern.NSeqPart(p) =>
         val sp = p match {
           case Wildcard => AnyElem #:: Lit('0') #:: Lit('1') #:: Stream.Empty
           case AnyElem => Lit('0') #:: Lit('1') #:: Stream.Empty
           case Lit(_) => Stream.Empty
         }
-        sp.map(Named.NPart(_))
-      case Named.NCat(fst, snd) =>
-        val s1 = shrinkNamed.shrink(fst)
-        val s2 = shrinkNamed.shrink(snd)
+        sp.map(NamedSeqPattern.NSeqPart(_))
+      case NamedSeqPattern.NCat(fst, snd) =>
+        val s1 = shrinkNamedSeqPattern.shrink(fst)
+        val s2 = shrinkNamedSeqPattern.shrink(snd)
         interleave(s1, s2).iterator.sliding(2).map {
-          case Seq(a, b) => Named.NCat(a, b)
-          case _ => Named.NEmpty
+          case Seq(a, b) => NamedSeqPattern.NCat(a, b)
+          case _ => NamedSeqPattern.NEmpty
         }
         .toStream
     }
 
-  def unany(p: Pattern): Pattern =
+  def unany[A](p: SeqPattern[A]): SeqPattern[A] =
     p match {
       case Empty => Empty
       case Cat(AnyElem, t) => unany(t)
       case Cat(h, t) => Cat(h, unany(t))
     }
 
-  def unwild(p: Pattern): Pattern =
+  def unwild[A](p: SeqPattern[A]): SeqPattern[A] =
     p match {
       case Empty => Empty
       case Cat(Wildcard, t) => unwild(t)
@@ -129,19 +139,24 @@ object StringSeqPatternGen {
 
 }
 
-abstract class SeqPatternLaws[E, S] extends FunSuite {
-  val seqPattern: SeqPattern.Aux[E, S]
+abstract class SeqPatternLaws[E, I, S, R] extends FunSuite {
+  import SeqPart.{Wildcard, AnyElem, Lit}
+
+  type Pattern = SeqPattern[E]
+  val Pattern = SeqPattern
+  type Named = NamedSeqPattern[E]
+  val Named = NamedSeqPattern
 
   implicit val generatorDrivenConfig =
     //PropertyCheckConfiguration(minSuccessful = 50000)
     PropertyCheckConfiguration(minSuccessful = 5000)
     //PropertyCheckConfiguration(minSuccessful = 5)
 
-  import seqPattern._
-
   def genPattern: Gen[Pattern]
   def genNamed: Gen[Named]
   def genSeq: Gen[S]
+  def splitter: Splitter[E, I, S, R]
+  def setOps: SetOps[Pattern]
 
   test("reverse is idempotent") {
     forAll(genPattern) { p: Pattern =>
@@ -149,10 +164,13 @@ abstract class SeqPatternLaws[E, S] extends FunSuite {
     }
   }
 
+  def matches(p: Pattern, s: S): Boolean
+  def namedMatches(p: Named, s: S): Boolean
+
   test("reverse matches the reverse string") {
-    forAll(genPattern, genSeq) { (p: Pattern, str: Sequence) =>
-      val rstr = listToSeq(seqToList(str).reverse)
-      assert(p.matches(str) == p.reverse.matches(rstr), s"p.reverse = ${p.reverse}")
+    forAll(genPattern, genSeq) { (p: Pattern, str: S) =>
+      val rstr = splitter.fromList(splitter.toList(str).reverse)
+      assert(matches(p, str) == matches(p.reverse, rstr), s"p.reverse = ${p.reverse}")
     }
   }
 
@@ -194,10 +212,10 @@ abstract class SeqPatternLaws[E, S] extends FunSuite {
     val p2 = Pattern.fromList(p20.toList.take(5))
     val n1 = p1.normalize
     val n2 = p2.normalize
-    val rawintr = p1.intersection(p2)
+    val rawintr = setOps.intersection(p1, p2)
     val intersect = rawintr.map(_.normalize).distinct
-    val sep = p1.matches(x) && p2.matches(x)
-    val together = intersect.exists(_.matches(x))
+    val sep = matches(p1, x) && matches(p2, x)
+    val together = intersect.exists(matches(_, x))
 
     assert(together == sep, s"n1: $n1, n2: $n2, intersection: $intersect")
     //if (together != sep) sys.error(s"n1: $n1, n2: $n2, intersection: $intersect")
@@ -210,29 +228,29 @@ abstract class SeqPatternLaws[E, S] extends FunSuite {
 
   test("matches any does match any") {
     forAll(genPattern, genSeq) { (p: Pattern, s: S) =>
-      if (p.matchesAny) assert(p.matches(s))
+      if (p.matchesAny) assert(matches(p, s))
     }
   }
 
   test("matchesEmpty is right") {
     forAll(genPattern) { p: Pattern =>
-      if (p.matchesEmpty) assert(p.matches(emptySeq))
+      if (p.matchesEmpty) assert(matches(p, splitter.emptySeq))
     }
   }
 
   test("toList/fromList match the same") {
     forAll(genPattern, genSeq) { (p, s) =>
       val p1 = Pattern.fromList(p.toList)
-      assert(p1.matches(s) == p.matches(s), s"p1 = $p1")
+      assert(matches(p1, s) == matches(p, s), s"p1 = $p1")
     }
   }
 
   test("isEmpty only matches empty") {
     forAll(genPattern, genSeq) { (p: Pattern, s: S) =>
       if (p.isEmpty) {
-        assert(p.matches(emptySeq))
-        if (s != emptySeq) {
-          assert(!p.matches(s))
+        assert(matches(p, splitter.emptySeq))
+        if (s != splitter.emptySeq) {
+          assert(!matches(p, s))
         }
       }
     }
@@ -240,18 +258,18 @@ abstract class SeqPatternLaws[E, S] extends FunSuite {
 
   test("unify union preserves matching") {
     forAll(Gen.listOf(genPattern), genSeq) { (ps: List[Pattern], s: S) =>
-      val u1 = ps.exists(_.matches(s))
-      val unified = seqPattern.Pattern.patternSetOps.unifyUnion(ps)
-      val u2 = unified.exists(_.matches(s))
+      val u1 = ps.exists(matches(_, s))
+      val unified = setOps.unifyUnion(ps)
+      val u2 = unified.exists(matches(_, s))
 
       assert(u2 == u1, s"unified = $unified")
     }
   }
 
   def subsetConsistentWithMatchLaw(a: Pattern, b: Pattern, s: S) = {
-    if (a.matches(s)) {
-      if (seqPattern.Pattern.patternSetOps.subset(a, b)) {
-        assert(b.matches(s))
+    if (matches(a, s)) {
+      if (setOps.subset(a, b)) {
+        assert(matches(b, s))
       }
     }
   }
@@ -266,17 +284,17 @@ abstract class SeqPatternLaws[E, S] extends FunSuite {
       val p1 = Pattern.fromList(p10.toList.take(5))
       val p2 = Pattern.fromList(p20.toList.take(5))
       // true difference is <= diff
-      val diff = seqPattern.Pattern.patternSetOps.difference(p1, p2)
-      val diffmatch = diff.exists(_.matches(s))
+      val diff = setOps.difference(p1, p2)
+      val diffmatch = diff.exists(matches(_, s))
 
-      if (diffmatch) assert(p1.matches(s), s"diff = $diff")
+      if (diffmatch) assert(matches(p1, s), s"diff = $diff")
 
-      if (p1.matches(s) && !p2.matches(s)) {
+      if (matches(p1, s) && !matches(p2, s)) {
         assert(diffmatch, s"diff = $diff")
       }
 
-      if (diff.isEmpty && p1.matches(s)) {
-        assert(p2.matches(s))
+      if (diff.isEmpty && matches(p1, s)) {
+        assert(matches(p2, s))
       }
 
       if (p2.matchesAny) assert(diff == Nil)
@@ -288,8 +306,8 @@ abstract class SeqPatternLaws[E, S] extends FunSuite {
 
   test("p + q match (s + t) if p.matches(s) && q.matches(t)") {
     forAll(genPattern, genPattern, genSeq, genSeq) { (p: Pattern, q: Pattern, s: S, t: S) =>
-      if (p.matches(s) && q.matches(t)) {
-        assert((p + q).matches(catSeq(s, t)))
+      if (matches(p, s) && matches(q, t)) {
+        assert(matches(p + q, splitter.catSeqs(s :: t :: Nil)))
       }
     }
   }
@@ -298,7 +316,7 @@ abstract class SeqPatternLaws[E, S] extends FunSuite {
   def namedMatchesPatternLaw(n: Named, str: S) = {
     val p = n.unname
 
-    assert(n.matches(str).isDefined == p.matches(str), s"${p.show}")
+    assert(namedMatches(n, str) == matches(p, str), s"${p.show}")
   }
 
   test("Named.matches agrees with Pattern.matches") {
@@ -319,9 +337,9 @@ abstract class SeqPatternLaws[E, S] extends FunSuite {
 */
 }
 
-
-class StringSeqPatternSetLaws extends SetOpsLaws[StringSeqPattern.Pattern] {
-  import StringSeqPattern._
+class StringSeqPatternSetLaws extends SetOpsLaws[SeqPattern[Char]] {
+  type Pattern = SeqPattern[Char]
+  val Pattern = SeqPattern
 
   implicit val generatorDrivenConfig =
     //PropertyCheckConfiguration(minSuccessful = 50000)
@@ -333,38 +351,60 @@ class StringSeqPatternSetLaws extends SetOpsLaws[StringSeqPattern.Pattern] {
     Pattern.fromList(p.toList.take(4)).normalize
   }
 
+  val pmatcher = Pattern.matcher(Splitter.stringSplitter(_ => ()))
+
+  def matches(p: Pattern, s: String): Boolean = pmatcher(p)(s).isDefined
+
   def eqUnion: Gen[Eq[List[Pattern]]] =
     Gen.listOfN(100, StringSeqPatternGen.genBitString).map { tests =>
       new Eq[List[Pattern]] {
         def eqv(a: List[Pattern], b: List[Pattern]) =
           tests.forall { s =>
-            a.exists(_.matches(s)) == b.exists(_.matches(s))
+            a.exists(matches(_, s)) == b.exists(matches(_, s))
           }
       }
     }
 
-  val setOps: SetOps[Pattern] = Pattern.patternSetOps
+  implicit val setOpsChar: SetOps[Char] = SetOps.distinct[Char]
+  val setOps: SetOps[Pattern] = Pattern.seqPatternSetOps[Char]
 }
 
-class SeqPatternTest extends SeqPatternLaws[Char, String] {
+class SeqPatternTest extends SeqPatternLaws[Char, Char, String, Unit] {
+  import SeqPart._
 
-  import StringSeqPattern._
-  import Pattern.patternSetOps._
-
-  lazy val seqPattern = StringSeqPattern
   def genPattern = StringSeqPatternGen.genPat
   def genNamed = StringSeqPatternGen.genNamed
   def genSeq = StringSeqPatternGen.genBitString
 
+  def splitter: Splitter[Char, Char, String, Unit] =
+    Splitter.stringSplitter(_ => ())
+
+  val pmatcher = Pattern.matcher(splitter)
+
+  def matches(p: Pattern, s: String): Boolean = pmatcher(p)(s).isDefined
+  def namedMatches(p: Named, s: String): Boolean =
+    NamedSeqPattern.matcher(splitter)(p)(s).isDefined
+
+  def namedMatch(p: Named, s: String): Option[Map[String, String]] =
+    NamedSeqPattern.matcher(Splitter.stringSplitter(_.toString))(p)(s)
+
+  implicit val setOpsChar: SetOps[Char] = SetOps.distinct[Char]
+  def setOps: SetOps[Pattern] = Pattern.seqPatternSetOps[Char]
+
   import StringSeqPatternGen._
 
+  def toPattern(s: String): Pattern =
+    s.toList.foldRight(Pattern.Empty: Pattern) { (c, r) =>
+      Pattern.Cat(Lit(c), r)
+    }
+
   test("some matching examples") {
-    val matches: List[(Pattern, String)] =
+   val ms: List[(Pattern, String)] =
       (Pattern.Wild + Pattern.Any + Pattern.Any + toPattern("1"), "111") ::
       (toPattern("1") + Pattern.Any + toPattern("1"), "111") ::
       Nil
 
-    matches.foreach { case (p, s) => assert(p.matches(s), s"$p.matches($s)") }
+    ms.foreach { case (p, s) => assert(matches(p, s), s"matches($p, $s)") }
   }
 
   test("wildcard on either side is the same as contains") {
@@ -393,16 +433,17 @@ class SeqPatternTest extends SeqPatternLaws[Char, String] {
   }
 
   test("subset is consistent with match regressions") {
-    assert(subset(toPattern("00") + Pattern.Wild, toPattern("0") + Pattern.Wild))
-    assert(subset(toPattern("00") + Pattern.Any + Pattern.Wild, toPattern("0") + Pattern.Any + Pattern.Wild))
+    assert(setOps.subset(toPattern("00") + Pattern.Wild, toPattern("0") + Pattern.Wild))
+    assert(setOps.subset(toPattern("00") + Pattern.Any + Pattern.Wild, toPattern("0") + Pattern.Any + Pattern.Wild))
   }
 
   test("if y - x is empty, (yz - xz) for all strings is empty") {
     def law(x0: Pattern, y0: Pattern, str: String) = {
       val x = Pattern.fromList(x0.toList.take(5))
       val y = Pattern.fromList(y0.toList.take(5))
-      if (y.difference(x) == Nil) {
-        assert((y + toPattern(str)).difference(x + toPattern(str)) == Nil)
+      if (setOps.difference(y, x) == Nil) {
+        val left = y + toPattern(str)
+        assert(setOps.difference(left, x + toPattern(str)) == Nil)
       }
     }
 
@@ -413,16 +454,17 @@ class SeqPatternTest extends SeqPatternLaws[Char, String] {
     forAll(genPattern, genPattern, genSeq) { (x0: Pattern, y0: Pattern, str: String) =>
       val x = Pattern.fromList(x0.toList.take(5))
       val y = Pattern.fromList(y0.toList.take(5))
-      if (y.difference(x) == Nil) {
-        assert((toPattern(str) + y).difference(toPattern(str) + x) == Nil)
+      if (setOps.difference(y, x) == Nil) {
+        val left = toPattern(str) + y
+        assert(setOps.difference(left, toPattern(str) + x) == Nil)
       }
     }
   }
 
   test("Named.matches + render agree") {
     forAll(genNamed, genSeq) { (n: Named, str: String) =>
-      n.matches(str).foreach { m =>
-        n.render(m)(c => c) match {
+      namedMatch(n, str).foreach { m =>
+        n.render(m)(_.toString) match {
           case Some(s0) => assert(s0 == str)
           case None =>
             // this can only happen if we have unnamed Wild/AnyElem
@@ -439,9 +481,9 @@ class SeqPatternTest extends SeqPatternLaws[Char, String] {
 
     import Named._
     val regressions: List[(Named, String)] =
-      (NCat(NEmpty,NCat(NPart(Lit('1')),NPart(Wildcard))), "1") ::
-      (NCat(NPart(Lit('1')),NPart(Wildcard)), "1") ::
-      (NPart(Lit('1')), "1") ::
+      (NCat(NEmpty,NCat(NSeqPart(Lit('1')),NSeqPart(Wildcard))), "1") ::
+      (NCat(NSeqPart(Lit('1')),NSeqPart(Wildcard)), "1") ::
+      (NSeqPart(Lit('1')), "1") ::
       Nil
 
     regressions.foreach { case (n, s) => namedMatchesPatternLaw(n, s) }
@@ -450,6 +492,6 @@ class SeqPatternTest extends SeqPatternLaws[Char, String] {
   test("Test some examples of Named matching") {
     // foo@("bar" baz@(*"baz"))
     val p1 = (named("bar") + (Named.Wild + named("baz")).name("baz")).name("foo")
-    assert(p1.matches("bar and baz") == Some(Map("foo" -> "bar and baz", "baz" -> " and baz")))
+    assert(namedMatch(p1, "bar and baz") == Some(Map("foo" -> "bar and baz", "baz" -> " and baz")))
   }
 }
